Add relocation of generic IA method to modules/vaccines
